requirements.txt
aiogram==2.21
alembic==1.8.1
asyncpg==0.26.0
psycopg2-binary==2.9.3
pydantic==1.9.1
SQLAlchemy==1.4.39
aiotelegraph==0.1.0
